Remove grafana direct integration - Fixes #489
