+++ horreum-backend/src/main/resources/application.properties
@@ -48,7 +48,7 @@ horreum.keycloak.url=http://localhost:8180
 
 # Address used in links to Horreum
 horreum.url=http://localhost:3000
-# Address used by internal services (e.g. Grafana) to access Horreum
+# Address used by internal services to access Horreum
 horreum.internal.url=http://localhost:8080
 
 horreum.alerting.missing.dataset.check=10m
@@ -62,13 +62,6 @@ horreum.transformationlog.max.lifespan=P30d
 quarkus.mailer.from=horreum@hyperfoil.io
 quarkus.mailer.host=localhost
 quarkus.mailer.port=2525
-
-horreum.grafana/mp-rest/url=http://localhost:4040
-horreum.grafana/mp-rest/scope=javax.inject.Singleton
-horreum.grafana.url=http://localhost:4040
-# On a production instance you should override this, too
-horreum.grafana.admin.user=admin
-horreum.grafana.admin.password=admin
 
 quarkus.package.type=fast-jar
 quarkus.container-image.registry=quay.io
