+++ horreum-backend/src/test/java/io/hyperfoil/tools/horreum/bus/MessageBusTestProfile.java
@@ -1,15 +1,15 @@
 package io.hyperfoil.tools.horreum.bus;
 
+import io.hyperfoil.tools.horreum.test.HorreumTestProfile;
+
 import java.util.HashMap;
 import java.util.Map;
 
-import io.hyperfoil.tools.horreum.test.NoGrafanaProfile;
-
-public class MessageBusTestProfile extends NoGrafanaProfile {
-   @Override
-   public Map<String, String> getConfigOverrides() {
-      Map<String, String> map = new HashMap<>(super.getConfigOverrides());
-      map.put("horreum.messagebus.retry.after", "1s");
-      return map;
-   }
+public class MessageBusTestProfile extends HorreumTestProfile {
+    @Override
+    public Map<String, String> getConfigOverrides() {
+        Map<String, String> map = new HashMap<>(super.getConfigOverrides());
+        map.put("horreum.messagebus.retry.after", "1s");
+        return map;
+    }
 }
