+++ horreum-client/src/main/java/io/hyperfoil/tools/HorreumClient.java
@@ -30,7 +30,7 @@ public class HorreumClient implements Closeable {
     public final AlertingService alertingService;
     public final BannerService bannerService;
     public final DatasetService datasetService;
-    public final GrafanaService grafanaService;
+    public final ChangesService changesService;
     public final NotificationService notificationService;
     public final ReportService reportService;
     public final RunServiceExtension runService;
@@ -41,14 +41,14 @@ public class HorreumClient implements Closeable {
     public final UserService userService;
 
     private HorreumClient(ResteasyClient client,
-                          ActionService actionService, AlertingService alertingService, BannerService bannerService, DatasetService datasetService, GrafanaService grafanaService,
+                          ActionService actionService, AlertingService alertingService, BannerService bannerService, DatasetService datasetService, ChangesService changesService,
                           NotificationService notificationService, ReportService reportService, RunServiceExtension runServiceExtension, SchemaService schemaService, SqlService sqlService,
                           SubscriptionService subscriptionService, TestService horreumTestService, UserService userService) {
         this.client = client;
         this.alertingService = alertingService;
         this.bannerService = bannerService;
         this.datasetService = datasetService;
-        this.grafanaService = grafanaService;
+        this.changesService = changesService;
         this.actionService = actionService;
         this.notificationService = notificationService;
         this.reportService = reportService;
@@ -177,7 +177,7 @@ public class HorreumClient implements Closeable {
                   target.proxyBuilder(AlertingService.class).build(),
                   target.proxyBuilder(BannerService.class).build(),
                   target.proxyBuilder(DatasetService.class).build(),
-                  target.proxyBuilder(GrafanaService.class).build(),
+                  target.proxyBuilder(ChangesService.class).build(),
                   target.proxyBuilder(NotificationService.class).build(),
                   target.proxyBuilder(ReportService.class).build(),
                   new RunServiceExtension(target, target.proxyBuilder(RunService.class).build()),
