+++ horreum-integration/src/test/resources/env.properties
@@ -2,7 +2,6 @@ horreum.username=user
 horreum.password=secret
 
 quarkus.datasource.password=secret
-horreum.grafana.admin.password=admin
 
 quarkus.test.enable-callbacks-for-integration-tests=true
