+++ horreum-model/src/main/java/io/hyperfoil/tools/horreum/changes/Target.java
@@ -1,4 +1,4 @@
-package io.hyperfoil.tools.horreum.grafana;
+package io.hyperfoil.tools.horreum.changes;
 
 public class Target {
    public String target;
