+++ horreum-model/src/main/java/io/hyperfoil/tools/horreum/changes/TimeRange.java
@@ -1,4 +1,4 @@
-package io.hyperfoil.tools.horreum.grafana;
+package io.hyperfoil.tools.horreum.changes;
 
 public class TimeRange {
    public String from = "now-30d";
