--- infra/grafana.env
@@ -1,16 +1,0 @@
-GF_INSTALL_PLUGINS=simpod-json-datasource
-GF_LOG_LEVEL=debug
-GF_SERVER_ROOT_URL=http://localhost:4040/
-GF_SERVER_HTTP_PORT=4040
-GF_SERVER_HTTP_ADDR=0.0.0.0
-GF_USERS_DEFAULT_THEME=light
-GF_SECURITY_ALLOW_EMBEDDING=true
-GF_AUTH_DISABLE_LOGIN_FORM=true
-GF_AUTH_OAUTH_AUTO_LOGIN=true
-GF_AUTH_GENERIC_OAUTH_ENABLED=true
-GF_AUTH_GENERIC_OAUTH_CLIENT_ID=grafana
-GF_AUTH_GENERIC_OAUTH_SCOPES=openid profile email
-GF_AUTH_GENERIC_OAUTH_ALLOW_SIGN_UP=false
-GF_AUTH_GENERIC_OAUTH_AUTH_URL=http://localhost:8180/realms/horreum/protocol/openid-connect/auth
-GF_AUTH_GENERIC_OAUTH_TOKEN_URL=http://localhost:8180/realms/horreum/protocol/openid-connect/token
-GF_AUTH_GENERIC_OAUTH_API_URL=http://localhost:8180/realms/horreum/protocol/openid-connect/userinfo
